add asseradasdds

+++ src/main/java/com/neu/group/controller/UserController.java
@@ -48,25 +48,30 @@ public class UserController {
     @PutMapping("/all")
     public R register(MultipartFile multiFile) throws IOException {
         String fileName = multiFile.getOriginalFilename();
-        String prefix = fileName.substring(fileName.lastIndexOf("."));
-        boolean flag = false;
-        File file = null;
-        try {
-            file = File.createTempFile(fileName, prefix);
-            multiFile.transferTo(file);
-            flag = userService.bulkImport(file);
-        } catch (Exception e) {
-            log.error(">>文件类型转换错误");
+        String prefix = null;
+        if (fileName != null) {
+            prefix = fileName.substring(fileName.lastIndexOf("."));
 
-        } finally {
-            // 操作完上面的文件 需要删除在根目录下生成的临时文件
-            assert file != null;
-            File f = new File(file.toURI());
-            Files.delete(f.toPath());
-            file.deleteOnExit();
+            boolean flag = false;
+            File file = null;
+            try {
+                file = File.createTempFile(fileName, prefix);
+                multiFile.transferTo(file);
+                flag = userService.bulkImport(file);
+            } catch (Exception e) {
+                log.error(">>文件类型转换错误");
+
+            } finally {
+                // 操作完上面的文件 需要删除在根目录下生成的临时文件
+                assert file != null;
+                File f = new File(file.toURI());
+                Files.delete(f.toPath());
+                file.deleteOnExit();
+            }
+
+            return new R(flag, "", flag ? "添加成功" : "添加失败,用户名重复");
         }
-
-        return new R(flag, "", flag ? "添加成功" : "添加失败,用户名重复");
+        return new R(false, "", "未导入文件");
     }
 
     //注销用户
